### 2016-12-09 #136 @whedon
[PRE REVIEW]: Limarka: making possible Brazilian students writing dissertation and thesis with Markdown
**Submitting author:** @edusantana (<a href="http://orcid.org/0000-0002-9165-9438">Eduardo Alexandre</a>)
**Repository:** <a href="https://github.com/abntex/limarka" target ="_blank">https://github.com/abntex/limarka</a>
**Version:** v0.5.2
