### 2016-12-09 #136 @acabunoc
Yes! I can take a look next week

> On Dec 9, 2016, at 1:52 AM, Arfon Smith <notifications@github.com> wrote:
>
> 👋 @acabunoc - would you be willing to edit this submission for JOSS?
>
