### 2016-12-10 #136 @edusantana
Thank you @acabunoc

I don't know if it's a problem, but the code, test, documentation, and examples are all writen in Portuguese.
