### 2016-12-14 2f889fa
inclusion of badge version
