### 2016-12-18 abntex/limarka#93 @edusantana
Several actions related to the software paper
I will sumarize here all the acitons related to [limarka's software paper](https://github.com/openjournals/joss-reviews/issues/136):

- [x]  Software paper writing #94
- [x] Creating an Open Canvas for limarka #95
- [x] Write a better CONTRIBUTING file #99
