+++ paper/open-canvas.md
@@ -1,0 +1,70 @@
+# Open Canvas
+
+**Project**: limarka
+
+# Problem
+
+<!-- The Top 1-3 problems you want to solve -->
+
+1. Saber as Normas da ABNT
+2. Saber como implementar as normas da ABNT no editor (Configuração de Páginas, seções, estilos, fomatação, sumário, apêndices, anexos, etc.)
+3. Manter consistente as citações e referências
+
+
+# Solution
+
+<!--  -->
+
+1. Exigir conhecimentos mínimos sobre as normas ABNT.
+2. Modelo com estilos pré-configurados e de fácil utilização.
+3. Gerenciamento de referências e sistema para citações.
+
+# Key Metrics
+
+<!-- How will you mesure success? -->
+
+- Estrelas no projeto
+- tráfego de acesso no github
+
+# Unique Value Proposition
+
+<!--  -->
+
+- Escrita do trabalho utilizando Markdown
+- Utilização de formulário PDF para configuração
+- Ótimo visual do PDF baseado no modelo Latex do abnTeX2
+
+# User profiles
+
+<!--  -->
+<!-- target audience and early adopters -->
+
+- Alunos de computação/informática que precisam escrever relatórios ou trabalhos de conclusão.
+
+# User channels
+
+<!--  -->
+
+- google groups
+- youtube channel
+- google plus channel
+
+# Resources required
+
+<!--  -->
+
+- Hardware: um computador pessoal.
+
+
+# Contributor Profiles
+
+<!-- Contributor types and ideal contributors -->
+
+- Latex expect
+- Ruby developer
+
+# Contributor Channels
+
+<!--  -->
+
+- Github issues
