### 2016-12-18 c3a423c
Creating an Open Canvas for limarka #95
