+++ paper/open-canvas.md
@@ -1,14 +1,16 @@
 # Open Canvas
 
-**Project**: limarka
+**Project**: [limarka](https://github.com/abntex/limarka)
+
+<!-- See http://mozillascience.github.io/working-open-workshop/writing_readme/ -->
 
 # Problem
 
 <!-- The Top 1-3 problems you want to solve -->
 
-1. Saber as Normas da ABNT
-2. Saber como implementar as normas da ABNT no editor (Configuração de Páginas, seções, estilos, fomatação, sumário, apêndices, anexos, etc.)
-3. Manter consistente as citações e referências
+1. Saber as Normas da ABNT atuais
+2. Saber como implementar as normas da ABNT atuais no editor (Configuração de Páginas, seções, estilos, fomatação, sumário, apêndices, anexos, etc.)
+3. Manter consistente as citações e referências no documento.
 
 
 # Solution
@@ -17,22 +19,26 @@
 
 1. Exigir conhecimentos mínimos sobre as normas ABNT.
 2. Modelo com estilos pré-configurados e de fácil utilização.
-3. Gerenciamento de referências e sistema para citações.
+3. Sistema para gerenciamento de citações e elaboração das referências automaticamente.
 
 # Key Metrics
 
 <!-- How will you mesure success? -->
 
-- Estrelas no projeto
-- tráfego de acesso no github
+- [Estrelas do projeto](https://github.com/abntex/limarka/stargazers)
+- Tráfegos de acesso no github: [limarka](https://github.com/abntex/limarka/graphs/traffic), [trabalho-academico-limarka
+](https://github.com/abntex/trabalho-academico-limarka
+/graphs/traffic)
+- [Notificações de trabalhos escritos com o limarka](https://github.com/abntex/limarka/wiki/M%C3%A9tricas)
 
 # Unique Value Proposition
 
 <!--  -->
 
-- Escrita do trabalho utilizando Markdown
-- Utilização de formulário PDF para configuração
-- Ótimo visual do PDF baseado no modelo Latex do abnTeX2
+- Escrita e formatação do documento fácil utilizando Markdown
+- Não exige conhecimentos técnicos para configuração, que é realizada através do preenchimento de um formulário PDF
+- Ótimo visual do documento de apresentação, pois é baseado no modelo Latex de trabalhos acadêmicos mantido pela comunidade abnTeX.
+- Gratuito, baseados em software livres.
 
 # User profiles
 
@@ -40,31 +46,44 @@
 <!-- target audience and early adopters -->
 
 - Alunos de computação/informática que precisam escrever relatórios ou trabalhos de conclusão.
+- Mestrandos em computação/informática que precisam escrever a dissertação. (*early adopters*)
 
 # User channels
 
 <!--  -->
 
-- google groups
-- youtube channel
-- google plus channel
+- Instruções das coordenações/instituições para produção dos trabalhos acadêmicos.
+- Descrição/instruções de tarefas submetidas aos alunos/aprendentes.
+- Páginas pessoais/institucionais de professores
+- Listas de discussões em cursos universitários
+- Notícias nas impressas universitárias
 
 # Resources required
 
 <!--  -->
 
 - Hardware: um computador pessoal.
-
+- Câmera e Microfone: gravação de vídeos para o canal.
 
 # Contributor Profiles
 
 <!-- Contributor types and ideal contributors -->
 
-- Latex expect
-- Ruby developer
+- Professores que solicitam trabalhos/relatórios aos alunos
+- Orientadores que desejam otimizar o tempo de discussões metodológicas com seus orientandos
+- Estudantes de informática que desejam produzir os trabalhos com agilidade
[…]
