### 2016-12-18 988d476
Creating an Open Canvas for limarka #95
