### 2016-12-19 d799927
Correção pra funcionar no Windows
No Windows não existe /dev/nul, devemos utilizar File::NULL para portabilidade

closes #96
