+++ lib/limarka/conversor.rb
@@ -52,7 +52,7 @@ module Limarka
     def compila
       Dir.chdir(options[:output_dir]) do
         basename = File.basename(texto_tex_file, '.tex')
-        system "latexmk --quiet --xelatex -f #{basename}", :out=>"/dev/null", :err=>"/dev/null"
+        system "latexmk --quiet --xelatex -f #{basename}",  :out=>File::NULL, :err=>File::NULL
         system "pdftotext -enc UTF-8 #{basename}.pdf"
         File.open("#{basename}.txt", 'r') {|f| @txt = f.read}
       end
