+++ lib/limarka/version.rb
@@ -1,3 +1,3 @@
 module Limarka
-  VERSION = "0.5.2"
+  VERSION = "0.5.3"
 end
