### 2016-12-19 72e15a5
Remover dependência de execução pdftotext closes #97
