### 2016-12-19 814ceb4
Remover dependência de execução pdftotext #97
