### 2016-12-19 6364bf6
Remover dependência de execução pdftotext #97
