+++ lib/limarka/conversor.rb
@@ -55,8 +55,10 @@ module Limarka
       Dir.chdir(options[:output_dir]) do
         basename = File.basename(texto_tex_file, '.tex')
         system "latexmk --quiet --xelatex -f #{basename}",  :out=>File::NULL, :err=>File::NULL
-        system "pdftotext -enc UTF-8 #{basename}.pdf"
-        File.open("#{basename}.txt", 'r') {|f| @txt = f.read}
+        if (usa_pdftotext) then
+          system "pdftotext -enc UTF-8 #{basename}.pdf"
+          File.open("#{basename}.txt", 'r') {|f| @txt = f.read}
+        end
       end
     end
 
