+++ CHANGELOG.md
@@ -1,11 +1,14 @@
 # Change Log
 
-## [v0.5.4](https://github.com/abntex/limarka/tree/v0.5.4) (2016-12-19)
-[Full Changelog](https://github.com/abntex/limarka/compare/v0.5.3...v0.5.4)
+## [v0.5.5](https://github.com/abntex/limarka/tree/v0.5.5) (2016-12-19)
+[Full Changelog](https://github.com/abntex/limarka/compare/v0.5.4...v0.5.5)
 
 **Melhorias implementadas:**
 
 - Remover dependência de execução pdftotext [\#97](https://github.com/abntex/limarka/issues/97)
+
+## [v0.5.4](https://github.com/abntex/limarka/tree/v0.5.4) (2016-12-19)
+[Full Changelog](https://github.com/abntex/limarka/compare/v0.5.3...v0.5.4)
 
 **Bugs corrigidos:**
 
