### 2016-12-20 7ed871d
Atualizando o README
