### 2016-12-21 25e9a92
Write a better CONTRIBUTING file #99
