### 2016-12-22 4da944e
Writing a better CONTRIBUTING file #99
