+++ CONTRIBUTING.md
@@ -11,16 +11,51 @@ O projeto tem como origem uma pesquisa científica (em andamento) sobre utiliza�
 
 # Professor
 
+Como professor você pode contribuir com o limarka divulgando-o e ofertando o limarka como uma opção para seus alunos realizarem tarefas. 
+
+- Ao solicitar relatórios de atividades aos alunos, sugira a utilização do limarka. Embora o limarka tenha o propósito de produzir trabalhos de conclusão de curso, ele pode ser facilmente utilizado para elaboração de relatórios. 
+
+- Sugira seus orientandos experimentarem o limarka. O limarka é gratuito e livre, pode ser utilizado no Linux, OS X e Windows. Não recomendamos utilizar o limarka em trabalhos de conclusão que exigam bastante fórmulas, teoremas ou provas. Os editores Latex oferecem melhor suporte para isso.
+
+- Atualize sua página pessoal/institucional adicionando uma [matéria ou link sobre o limarka](https://github.com/abntex/limarka/wiki/Imprensa). Seus alunos e outras pessoas poderão conhecer a ferramenta dessa forma.
+
+- Solicite incluir o limarka como uma alternativa nos materiais de instruções de produção de trabalhos de conclusão, ou nas disciplinas de Metodologia. O momento ideal para apresentar aos alunos a ferramenta são nas aulas de Metodologia. Por enquanto o foco da ferramenta são trabalho de conclusão, mas em breve também será possível realizar artigos.
+
+- Divulgue um notícia sobre o limarka nas listas de professores ou alunos. O público alvo da ferramenta são os estudantes, essas listas propocionam encontrá-los.
+
+- Recomende a veiculação de uma matéria sobre o limarka na agência de noticias de sua instituição. As agências de notícias das instituições são um ótimo meio para divulgação, verifique como contactar agência e recomende uma matéria para ser veiculada.
+
+- Clique na estrela [no repositório do projeto](https://github.com/abntex/limarka) (equivale a um *like*) no github.
+
 # Estudante
+
+- Utilize o limarka para elaboração de relatórios
+- Participe do [experimente de uso da ferramenta](https://github.com/abntex/limarka/wiki/Experimentos) - NECESSITA-SE URGENTEMENTE!
+- Compartilhe alguma notícia do limarka em suas redes sociais ou lista de alunos ou de laboratórios
+- Escreva seu trabalho de conclusão ou sua proposta/projeto com o limarka e compartilhe sua experiência
+- Clique na estrela [no repositório do projeto](https://github.com/abntex/limarka) (equivale a um *like*) no github.
 
 # Designer
 
+- [Precisamos de ajuda para fazer um logotipo para o projeto](https://github.com/abntex/limarka/issues/102).
+
+# Usuário Latex
+
+TODO.
+
 # Desenvolvedor Ruby
 
-Pull requests são sempre bem vindos. Ao participar desse projeto você está de 
-acordo com o [código de conduda].
+Deseja participar do desenvolvimento do limarka? Considere os [issues marcados com o label help needed](https://github.com/abntex/limarka/labels/help%20wanted) como um convite para contribuir conosco.
+
+Correção de bugs e implementação de novas funcionalidades através de Pull requests são sempre bem vindos.
+
+Ao participar desse projeto você está de acordo com o [código de conduda].
 
 [código de conduda]: https://github.com/abntex/limarka/blob/master/CODE_OF_CONDUCT.md
+
+## Implementando uma nova funcionalidade
+
+[Crie um issue](https://github.com/abntex/limarka/issues/new) descrevendo a nova funcionalidade.
 
[…]
