### 2016-12-23 13a9956
Adição de logo #102
