### 2016-12-23 9dee5b2
Adição de logo #102
