### 2016-12-23 d8dfcc8
Adição de uso para relatório
