### 2016-12-27 7c4adce
Inclusão de menu interativo
A ideia é que esse menu possa ser executado através de um bat e facilite os usuários do windows que não estão acostumados com comandos de linha.

closes #100
