### 2017-01-05 1e62aa2
Inclusão de badge para o gitter closes #106
