### 2017-01-05 00ff527
Inclusão de badge para o gitter closes #106
