### 2017-01-06 c962b7c
Indicação de autocompletar no menu interativo
