### 2017-01-06 9fb8a3e
Utilizar sistema de integração contínua #108
- Utilizando sudo para instalar dependências

- Ignorando teste para passar no travis

- Atualizando versão do pandoc

- Instalando dependências completas do ruby, talvez nao seja necessário.

- Adicionando libreoffice para execução dos testes
