### 2017-01-06 b2637da
Inclusão do badge do build closes #108
