### 2017-01-06 f51da3f
Fazer os testes passarem no servidor de integração contínua #109
- utilizando ubuntu 14.04 como versão mínima
- utilizando pandoc da distribuição
