### 2017-01-07 9082b0a
Fazer os testes passarem no servidor de integração contínua #109
Executando os testes sem latex
