+++ spec/lista_ilustracoes_spec.rb
@@ -22,7 +22,7 @@ TEXTO
     FileUtils.mkdir_p output_dir
   end
 
-  context 'quando ativada',  :pdf, :lento do
+  context 'quando ativada',  :compilacao, :lento do
     let (:output_dir) {"tmp/lista_ilustracoes/ativada"}
     let (:configuracao_especifica) {{'lista_ilustracoes' => true}}
     let (:t) {Limarka::Trabalho.new(configuracao: configuracao_padrao.merge(configuracao_especifica), texto: texto)}
@@ -39,7 +39,7 @@ TEXTO
     end
   end
 
-  context 'quando desativada',  :pdf, :lento do
+  context 'quando desativada',  :compilacao, :lento do
     let (:output_dir) {"tmp/lista_tabelas/desativada"}
     let (:configuracao_da_tabela) {{'lista_tabelas' => false}}
     let (:t) {Limarka::Trabalho.new(configuracao: configuracao_padrao.merge(configuracao_da_tabela), texto: texto)}
