### 2017-01-08 #136 @acabunoc
Sorry for the delay here!

@rgaiacs, would you be interested in being a reviewer for JOSS and reviewing this paper? Review guidelines can be found here: http://joss.theoj.org/about#reviewer_guidelines

Also, @edusantana do you have any potential reviewers you can suggest (someone close to the domain & speaks Portuguese)? It's nice to have multiple reviewers, and I'm not sure which of current reviewers knows Portuguese.
