### 2017-01-08 #136 @edusantana
Since limarka is intended to make possible to write scientific work (according to Brazilian's rules) I think these reviewers would be interested:

@jsta : because of reproducible science and scientific writing tools
@richardlitt : because of computer science, computational linguistics, data science

@acabunoc I can deliver translations to English of important parts of documentation.
