### 2017-01-08 #136 @edusantana
I do have a question, can I have the paper on a different repository just for it? I put it on a different branch, but that doesn't look good to me.
