+++ spec/latex/exemplo-pequeno-latex/xxx-referencias.bib
@@ -1,0 +1,369 @@
+@article{almeida2013,
+  title = {Investigação sobre os Fatores de Risco da Prematuridade},
+  author = {de Almeida, Thassiany Sarmento Oliveira and Lins, Raquel Pinto and de Lima Camelo, Ademir and de Mello, Dandarah Christie Cavalcanti Lima},
+  journal = {Revista Brasileira de Ciências da Saúde},
+  year = {2013},
+  issn = {1415-2177},
+  number = {3},
+  pages = {301-308},
+  volume = {17},
+  doi = {10.4034/RBCS.2013.17.03.14},
+  owner = {eduardo},
+  timestamp = {2016.11.20},
+  url = {http://periodicos.ufpb.br/index.php/rbcs/article/view/13674/9814},
+  urlaccessdate = {18 de nov. de 2016},
+  subtitle = {uma Revisão Sistemática}
+}
+@manual{abntex2cite,
+  title = {O pacote abntex2cite},
+  author = {Araujo, Lauro C{\'e}sar},
+  organization = {Equipe abnTeX2},
+  year = {2015},
+  annote = {Este documento {\'e} derivado do \cite{abnt-bibtex-doc}},
+  bdsk-url-1 = {http://www.abntex.net.br/},
+  date-added = {2013-01-09 10:37:45 +0000},
+  date-modified = {2015-04-27 22:43:38 +0000},
+  url = {http://www.abntex.net.br/},
+  subtitle = {Estilos bibliogr{\'a}ficos compat{\'\i}veis com a ABNT NBR 6023}
+}
+@manual{abntex2cite-alf,
+  title = {O pacote abntex2cite},
+  author = {Araujo, Lauro C{\'e}sar},
+  organization = {Equipe abnTeX2},
+  year = {2015},
+  annote = {Este documento {\'e} derivado do \cite{abnt-bibtex-alf-doc}},
+  bdsk-url-1 = {http://www.abntex.net.br/},
+  date-added = {2013-01-09 10:37:45 +0000},
+  date-modified = {2015-04-27 22:43:44 +0000},
+  url = {http://www.abntex.net.br/},
+  subtitle = {t{\'o}picos espec{\'\i}ficos da ABNT NBR 10520}
+}
+@manual{abntex2classe,
+  title = {A classe abntex2},
+  author = {Araujo, Lauro C{\'e}sar},
+  organization = {Equipe abnTeX2},
+  year = {2015},
+  bdsk-url-1 = {http://www.abntex.net.br/},
+  date-added = {2013-01-09 10:37:38 +0000},
+  date-modified = {2015-04-27 22:42:47 +0000},
+  url = {http://www.abntex.net.br/},
+  subtitle = {Modelo can{\^o}nico de trabalhos acad{\^e}micos brasileiros compat{\'\i}vel com as normas ABNT NBR 14724}
+}
+@manual{abntex2modelo,
+  title = {Modelo Can{\^o}nico de Trabalho Acad{\^e}mico com abnTeX2},
+  author = {Araujo, Lauro C{\'e}sar},
+  organization = {Equipe abnTeX2},
+  year = {2015},
+  annote = {Este documento {\'e} derivado do \cite{abnt-bibtex-doc}},
+  bdsk-url-1 = {http://www.abntex.net.br/},
+  date-added = {2013-01-12 22:55:32 +0000},
+  date-modified = {2015-04-27 22:43:32 +0000},
+  url = {http://www.abntex.net.br/}
+}
+@manual{abntex2modelo-artigo,
+  title = {Modelo Can{\^o}nico de Artigo Cient{\'\i}fico com abnTeX2},
+  author = {Araujo, Lauro C{\'e}sar},
+  organization = {Equipe abnTeX2},
[…]
