### 2017-01-08 5a63eb1
Testes padrão excluem latex #109
Um branch separada para os testes de compilação latex foi preparada.
Enquanto ela não estiver passando os testes de latex foram removidos da master.

Utilizando o script de test padrão do travis: `rake`
