+++ .travis.yml
@@ -9,4 +9,4 @@ before_install:
 - wget https://github.com/jgm/pandoc/releases/download/1.19.1/pandoc-1.19.1-1-amd64.deb
 - sudo dpkg -i pandoc-1.19.1-1-amd64.deb
 - sudo apt-get install libreoffice
-script: bundle exec rake configuracao_padrao spec:semlatex
+# Utilizando o script default do travis
