### 2017-01-08 734f74e
Configurar deployment a partir do travis closes #111
