+++ spec/dependencias/latex_spec.rb
@@ -3,19 +3,41 @@
 require 'spec_helper'
 require 'limarka/conversor'
 
-describe 'Latex dependências', :dependencias do
-  
-  let(:input_dir) {"spec/latex/exemplo-pequeno-latex"}
-  let(:output_dir) {"test/latex/exemplo-pequeno-latex"}
-  let(:tex_file) {"xxx-Monografia.tex"}
+describe 'Compilação Latex', :dependencias, :dependencias_latex do
 
+  let(:output_dir) {input_dir.gsub("spec","test")}
+  let(:tex_file) {"xxx-Monografia.tex"}
+  
   before do
     FileUtils.rm_rf output_dir
     FileUtils.mkdir_p output_dir
     FileUtils.cp_r input_dir+"/.", output_dir
   end
 
+  context 'de arquivo abntex2 mínimo com latexmk', :latex_minimo do
+    let(:input_dir) {"spec/latex/exemplo-minimo"}
+    
+    before do
+      Dir.chdir(output_dir) do
+        system "latexmk -quiet -pdflatex=\"xelatex %O %S\" -pdf -dvi- -ps- -f  xxx-Monografia.tex"
+        system "pdftotext -enc UTF-8 xxx-Monografia.pdf"
+      end
+    end
+    
+    it 'gera o PDF' do
+      expect(File).to exist(output_dir+"/xxx-Monografia.pdf")
+    end
+
+    it 'o capítulo é prefixado com o número dele' do
+      expect(File.read(output_dir+"/xxx-Monografia.txt")).to include("1 Chapter example")
+    end
+
+    
+  end
+
+
   context 'compilando um arquivo tex pequeno com latexmk' do
+    let(:input_dir) {"spec/latex/exemplo-pequeno-latex"}
    
     it 'gera o PDF' do
       Dir.chdir(output_dir) do
