+++ .travis.yml
@@ -6,6 +6,7 @@ rvm:
 before_install:
 - sudo apt-get update -qq
 - sudo apt-get install -qq ruby-full pdftk poppler-utils xclip latexmk texlive texlive-publishers texlive-lang-portuguese texlive-latex-extra texlive-fonts-recommended texlive-xetex
+- sudo apt-get install -qq language-pack-pt
 - wget https://github.com/jgm/pandoc/releases/download/1.19.1/pandoc-1.19.1-1-amd64.deb
 - sudo dpkg -i pandoc-1.19.1-1-amd64.deb
 #- sudo apt-get install libreoffice
