### 2017-01-08 fd5b86e
instalando texlive através da rede
instruções podem ser vistas em http://www.tug.org/texlive/doc/texlive-en/texlive-en.html#installation
