+++ .travis.yml
@@ -5,10 +5,13 @@ rvm:
 - 2.3.1
 before_install:
 - sudo apt-get update -qq
-- sudo apt-get install -qq ruby-full pdftk poppler-utils xclip latexmk texlive texlive-publishers texlive-lang-portuguese texlive-latex-extra texlive-fonts-recommended texlive-xetex
+- sudo apt-get install -qq ruby-full pdftk poppler-utils xclip
 - sudo apt-get install -qq language-pack-pt
 - wget https://github.com/jgm/pandoc/releases/download/1.19.1/pandoc-1.19.1-1-amd64.deb
 - sudo dpkg -i pandoc-1.19.1-1-amd64.deb
+- wget http://mirror.ctan.org/systems/texlive/tlnet/install-tl-unx.tar.gz
+- cd install-tl-* && ./install-tl -no-gui -scheme full
 #- sudo apt-get install libreoffice
 script: 
 - bundle exec rspec -t latex_minimo
+
