### 2017-01-08 4c0e85c
Descompactando antes de executar instalação
