+++ .travis.yml
@@ -10,7 +10,7 @@ before_install:
 - wget https://github.com/jgm/pandoc/releases/download/1.19.1/pandoc-1.19.1-1-amd64.deb
 - sudo dpkg -i pandoc-1.19.1-1-amd64.deb
 - wget http://mirror.ctan.org/systems/texlive/tlnet/install-tl-unx.tar.gz
-- cd install-tl-* && ./install-tl -no-gui -scheme full
+- tar -xf install-tl-unx.tar.gz && cd install-tl-* && ./install-tl -no-gui -scheme full
 #- sudo apt-get install libreoffice
 script: 
 - bundle exec rspec -t latex_minimo
