### 2017-01-08 25bb6fa
Adicionando permissão de root na instalação
