+++ .travis.yml
@@ -10,7 +10,8 @@ before_install:
 - wget https://github.com/jgm/pandoc/releases/download/1.19.1/pandoc-1.19.1-1-amd64.deb
 - sudo dpkg -i pandoc-1.19.1-1-amd64.deb
 - wget http://mirror.ctan.org/systems/texlive/tlnet/install-tl-unx.tar.gz
-- tar -xf install-tl-unx.tar.gz && cd install-tl-* && sudo ./install-tl -no-gui -scheme full
+- tar -xf install-tl-unx.tar.gz && cd install-tl-*
+- echo 'selected_scheme scheme-small' > temp.profile && sudo ./install-tl -no-gui -profile temp.profile
 #- sudo apt-get install libreoffice
 script: 
 - bundle exec rspec -t latex_minimo
