### 2017-01-08 3d39975
Executando todas os testes latex
