### 2017-01-08 9eba4ff
latex_minimo
