### 2017-01-09 #136 @acabunoc
Thanks @edusantana! If you're willing to provide an English translation, that would help a lot for finding reviewers. If @jsta, @RichardLitt or @rgaiacs are interested, reply with a quick ✋  -- otherwise, feel free to ignore this issue :)

For the paper, it's usually on the master branch in the main repo. Sometimes it's on a permanent branch based off the reviewed version. I'll defer to @arfon on whether or not having the paper in a completely separate repo is okay.
