### 2017-01-09 edd6175
Atualizando do changelog
