### 2017-01-09 9dbb3bf
Remoção de dependências desnecessárias e atualização para schema-full
