+++ .travis.yml
@@ -5,13 +5,12 @@ rvm:
 - 2.3.1
 before_install:
 - sudo apt-get update -qq
-- sudo apt-get install -qq ruby-full pdftk poppler-utils xclip latexmk libreoffice
+- sudo apt-get install -qq poppler-utils
 - sudo apt-get install -qq language-pack-pt
-- wget https://github.com/jgm/pandoc/releases/download/1.19.1/pandoc-1.19.1-1-amd64.deb
-- sudo dpkg -i pandoc-1.19.1-1-amd64.deb
 - wget http://mirror.ctan.org/systems/texlive/tlnet/install-tl-unx.tar.gz
-- tar -xf install-tl-unx.tar.gz && cd install-tl-*
-- echo 'selected_scheme scheme-small' > temp.profile && sudo ./install-tl -no-gui -profile temp.profile && cd ..
-script: 
+- tar -xf install-tl-unx.tar.gz
+- cd install-tl-* && echo 'selected_scheme scheme-full' > temp.profile && sudo ./install-tl -no-gui -profile temp.profile && cd ..
+script:
+- latexmk -v
+- xelatex -v
 - bundle exec rspec -t latex_minimo
-
