### 2017-01-09 5fd9884
Criação de profile de instalação e adição ao PATH
Adição do texlive ao path

Fazer os testes Latex passarem no travis #113
