+++ .travis.yml
@@ -5,12 +5,14 @@ rvm:
 - 2.3.1
 before_install:
 - sudo apt-get update -qq
-- sudo apt-get install -qq poppler-utils
-- sudo apt-get install -qq language-pack-pt
+- sudo apt-get install -y -qq poppler-utils language-pack-pt
+install:
 - wget http://mirror.ctan.org/systems/texlive/tlnet/install-tl-unx.tar.gz
 - tar -xf install-tl-unx.tar.gz
-- cd install-tl-* && echo 'selected_scheme scheme-full' > temp.profile && sudo ./install-tl -no-gui -profile temp.profile && cd ..
+- cd install-tl-* && && sudo ./install-tl -no-gui -profile ./spec/dependencias/texlive-abntex2.profile && cd ..
+- export PATH=$PATH:/usr/local/texlive/2016/bin/x86_64-linux
 script:
 - latexmk -v
 - xelatex -v
 - bundle exec rspec -t latex_minimo
+
