### 2017-01-10 70d1f24
Instalação mínima
