+++ .travis.yml
@@ -9,7 +9,7 @@ before_install:
 install:
 - wget http://mirror.ctan.org/systems/texlive/tlnet/install-tl-unx.tar.gz
 - tar -xf install-tl-unx.tar.gz
-- cd install-tl-* && sudo ./install-tl -no-gui -profile ./spec/dependencias/texlive-abntex2.profile && cd ..
+- cd install-tl-* && sudo ./install-tl -no-gui -profile ./spec/dependencias/texlive-abntex2.profile -no-verify-downloads && cd ..
 - export PATH=$PATH:/usr/local/texlive/2016/bin/x86_64-linux
 script:
 - latexmk -v
