### 2017-01-10 9534730
Instalação mínima
