### 2017-01-10 3d2dd3c
inclusão de ruby e latexmk
