### 2017-01-10 #136 @arfon
> For the paper, it's usually on the master branch in the main repo. Sometimes it's on a permanent branch based off the reviewed version. I'll defer to @arfon on whether or not having the paper in a completely separate repo is okay.

@edusantana - the paper needs to be in the repository being reviewed. Is there an reason this can't be in a `docs` or `paper` folder on the `master` branch?
