+++ .travis.yml
@@ -14,4 +14,6 @@ script:
 - latexmk -v
 - xelatex -v
 - bundle exec rspec -t latex_minimo
+after_failure:
+- cd /usr/local/texlive && find . -iname latexmk
 
