### 2017-01-10 3bfc046
Correção dos testes
