+++ spec/dependencias/latex_spec.rb
@@ -4,40 +4,27 @@ require 'spec_helper'
 require 'limarka/conversor'
 
 describe 'Compilação Latex', :dependencias, :dependencias_latex do
-
-
-  let(:tex_file) {"xxx-Monografia.tex"}
   
   before do
+    # Cria cópia de input_dir em output_dir
     FileUtils.rm_rf output_dir
     FileUtils.mkdir_p output_dir
     FileUtils.cp_r input_dir+"/.", output_dir
   end
 
-  
-
   context 'de arquivo abntex2 mínimo com latexmk', :latex_minimo do
     let(:input_dir) {"spec/latex/exemplo-minimo"}
     let(:output_dir) {input_dir.gsub("spec","test")}
     
-    before do
+    it 'O pdf é gerado e o capítulo é prefixado com o número dele' do
       Dir.chdir(output_dir) do
-        system "latexmk -quiet -pdflatex=\"xelatex %O %S\" -pdf -dvi- -ps- -f  xxx-Monografia.tex"
-      end
-    end
-    
-    it 'gera o PDF' do
-      expect(File).to exist(output_dir+"/xxx-Monografia.pdf")
-    end
-
-    it 'o capítulo é prefixado com o número dele' do
-      expect(File).to exist(output_dir+"/xxx-Monografia.pdf")
-      system "pdftotext -enc UTF-8 xxx-Monografia.pdf"
-      expect(File).to exist(output_dir+"/xxx-Monografia.txt")
-      expect(File.read(output_dir+"/xxx-Monografia.txt")).to include("1 Chapter example")
-    end
-
-    
+        system "latexmk --quiet --xelatex -f xxx-Monografia.tex"
+        expect(File).to exist("xxx-Monografia.pdf")
+        system "pdftotext -enc UTF-8 xxx-Monografia.pdf"
+        expect(File).to exist("xxx-Monografia.txt")
+        expect(File.read("xxx-Monografia.txt")).to include("1 Chapter example")
+      end    
+    end    
   end
 
 
@@ -45,24 +32,15 @@ describe 'Compilação Latex', :dependencias, :dependencias_latex do
     let(:input_dir) {"spec/latex/exemplo-minimo-com-acentos"}
     let(:output_dir) {input_dir.gsub("spec","test")}
     
-    before do
+    it 'O pdf é gerado e o capítulo é prefixado com o número dele' do
       Dir.chdir(output_dir) do
-        system "latexmk -quiet -pdflatex=\"xelatex %O %S\" -pdf -dvi- -ps- -f  xxx-Monografia.tex"
-      end
-    end
-    
-    it 'gera o PDF' do
-      expect(File).to exist(output_dir+"/xxx-Monografia.pdf")
-    end
-
-    it 'o capítulo é prefixado com o número dele' do
-      expect(File).to exist(output_dir+"/xxx-Monografia.pdf")
-      system "pdftotext -enc UTF-8 xxx-Monografia.pdf"
-      expect(File).to exist(output_dir+"/xxx-Monografia.txt")
-      expect(File.read(output_dir+"/xxx-Monografia.txt")).to include("1 Introdução")
-    end
-
-    
+        system "latexmk --quiet --xelatex -f xxx-Monografia.tex"
+        expect(File).to exist("xxx-Monografia.pdf")
+        system "pdftotext -enc UTF-8 xxx-Monografia.pdf"
+        expect(File).to exist("xxx-Monografia.txt")
+        expect(File.read("xxx-Monografia.txt")).to include("1 Introdução")
+      end    
+    end   
   end
 
 
@@ -70,24 +48,15 @@ describe 'Compilação Latex', :dependencias, :dependencias_latex do
[…]
