### 2017-01-10 cf42ae5
Buscando instalação mínima do latex
