### 2017-01-10 0542039
Retornando para full install
