### 2017-01-11 #136 @edusantana
@arfon  I did not have a `docs` folder, I use wiki for documentation, but I will create one now. And the paper will be there.
