### 2017-01-11 281f714
Bring paper to master #93
