+++ paper/paper.bib
@@ -1,40 +1,0 @@
-@online{Pandoc,
-  author = {John MacFarlane},
-  title = {Pandoc a universal document converter},
-  year = 2007,
-  url = {http://pandoc.org},
-  urldate = {2016-12-08}
-}
-
-@online{abnTeX2,
-  author = {abnTeX},
-  title = {Classe LaTeX e estilo bibliográfico compatíveis com as normas da ABNT},
-  year = 2016,
-  url = {http://www.abntex.net.br},
-  urldate = {2016-12-08}
-}
-
-@online{YAML,
-  author = {YAML},
-  title = {YAML Ain't Markup Language},
-  year = 2001,
-  url = {http://yaml.org},
-  urldate = {2016-12-08}
-}
-
-
-@online{form,
-  author = {Richard Foersom},
-  title = {Create PDF Forms with LibreOffice},
-  year = 2016,
-  url = {http://foersom.com/org/HowTo/CreatePdfForm.html},
-  urldate = {2016-12-08}
-}
-
-@manual{NBR14724:2011,
-	Address = {Rio de Janeiro},
-	Org-Short = {ABNT},
-	Organization = {Associa{\c c}\~ao Brasileira de Normas T\'ecnicas},
-	Pages = 15,
-	Title = {{NBR} 14724},
-  Year = 2011}
