+++ paper/open-canvas.md
@@ -1,86 +1,0 @@
-# Open Canvas
-
-**Project**: [limarka](https://github.com/abntex/limarka)
-
-<!-- See http://mozillascience.github.io/working-open-workshop/writing_readme/ -->
-
-# Problem
-
-<!-- The Top 1-3 problems you want to solve -->
-
-1. Saber as Normas da ABNT atuais
-2. Saber como implementar as normas da ABNT atuais no editor (Configuração de Páginas, seções, estilos, formatação, sumário, apêndices, anexos, etc.)
-3. Manter consistente as citações e referências no documento.
-
-
-# Solution
-
-<!--  -->
-
-1. Exigir conhecimentos mínimos sobre as normas ABNT.
-2. Modelo com estilos pré-configurados e de fácil utilização.
-3. Sistema para gerenciamento de citações e elaboração das referências automaticamente.
-
-# Key Metrics
-
-<!-- How will you mesure success? -->
-
-- [Estrelas do projeto](https://github.com/abntex/limarka/stargazers)
-- Tráfegos de acesso no github: [limarka](https://github.com/abntex/limarka/graphs/traffic), [trabalho-academico-limarka
-](https://github.com/abntex/trabalho-academico-limarka
-/graphs/traffic)
-- [Notificações de trabalhos escritos com o limarka](https://github.com/abntex/limarka/wiki/M%C3%A9tricas)
-
-# Unique Value Proposition
-
-<!--  -->
-
-- O documento é escrito e formatado facilmente utilizando marcações no texto
-- Não exige conhecimentos técnicos para configuração, que é realizada através do preenchimento de um formulário PDF
-- Ótimo visual do documento de apresentação, pois é baseado no modelo Latex de trabalhos acadêmicos mantido pela comunidade abnTeX.
-- Solução Gratuita, software livre.
-
-# User profiles
-
-<!--  -->
-<!-- target audience and early adopters -->
-
-- Alunos com conhecimentos técnicos de informática que precisam escrever relatórios ou trabalhos de conclusão.
-- Mestrandos em computação/informática que precisam escrever a dissertação. (*early adopters*)
-
-# User channels
-
-<!--  -->
-
-- Instruções das coordenações/instituições para produção dos trabalhos acadêmicos sugerindo utilização do limarka.
-- Instruções de tarefas submetidas aos alunos/aprendentes sugerindo utilização do limarka.
-- Páginas pessoais/institucionais de professores.
-- Listas de discussões de alunos universitários.
-- Publicidade nas agências de noticias universitárias.
-
-# Resources required
-
-<!--  -->
-- Sistema para conversão de Markdown para Latex;
-- Templates Latex para produção de trabalhos de conclusão em conformidade com as Normas da ABNT;
-- Desenvolvedor ruby;
-- Estrutural inicial de trabalho de conclusão de curso;
-- Câmera e Microfone: gravação de vídeos para o canal.
-
-# Contributor Profiles
-
-<!-- Contributor types and ideal contributors -->
-
-- Professores que solicitam trabalhos/relatórios aos alunos
-- Orientadores que desejam otimizar o tempo de discussões metodológicas com seus orientandos
-- Estudantes de informática que desejam produzir os trabalhos com agilidade
-- Usuários Latex que desejam elaborar novos templates
-- Desenvolvedores Ruby que desejam implementar novas funcionalidades
-
-# Contributor Channels
-
-<!--  -->
-
-- [Publicidade nas agências de noticias universitárias](https://github.com/abntex/limarka/wiki/Imprensa)
-- [CONTRIBUTING.md no repositório do projeto](https://github.com/abntex/limarka/blob/master/CONTRIBUTING.md)
-- [Canal do limarka Youtube](https://www.youtube.com/channel/UC4kERewtyT_nw__YHbKMdBA)
