### 2017-01-12 23cbd2f
Remoção de parte duplicada do README
