+++ README.md
@@ -119,124 +119,4 @@ O limarka converte os textos produzidos em Markdown para LaTeX, utilizando os mo
 # License
 
 Esse gem é disponibilizado como software livre através da [Licença MIT](http://opensource.org/licenses/MIT).
-![logo](https://raw.githubusercontent.com/wiki/abntex/limarka/imagens/limarka-logo.png)
 
-[![Build Status](https://travis-ci.org/abntex/limarka.svg?branch=master)](https://travis-ci.org/abntex/limarka) [![Gem Version](https://badge.fury.io/rb/limarka.svg)](https://badge.fury.io/rb/limarka) [![Join the chat at https://gitter.im/abntex/limarka](https://badges.gitter.im/abntex/limarka.svg)](https://gitter.im/abntex/limarka?utm_source=badge&utm_medium=badge&utm_campaign=pr-badge&utm_content=badge)
-
-![Normas da ABNT em 2017: Atualizadas](https://img.shields.io/badge/Normas da ABNT em 2017-Atualizadas-brightgreen.svg) ![NBR 14724:2011 100%](https://img.shields.io/badge/NBR 2014724%3A2011-100%25-brightgreen.svg)
-
-Olá! Bem vindo ao limarka!
-
-<!-- Show what you're doing, for who, and why.  -->
-
-O limarka é uma ferramenta que possibilita seus usuários escreverem relatórios ou trabalhos de conclusão de curso (Monografia, TCC, Dissertação ou Tese) em arquivos somente texto e **produzir PDFs em conformidades com as Normas da ABNT**. A produção desse tipo de trabalho é um momento sensível na vida dos estudantes, que pode contribuir para evasão no ensino superior. O objetivo do limarka é liberar dos estudantes os esforços de formatação e adequação às Normas da ABNT, deixando-os concentrar-se no planejamento, execução e escrita de seus trabalhos.
-
-
-# Especificidades do limarka
-
-<!-- Explain what makes your project special, useful, exciting! -->
-
-* Exige conhecimentos mínimos sobre as Normas da ABNT
-* Utiliza modelo com estilos pré-configurados com as normas e de fácil utilização
-* Utiliza sistema para gerenciamento de citações e elaboração das referências automaticamente
-* Gera o sumário automaticamente, sem esforço adicional
-* Permite a escrita do trabalho em arquivos somente texto, utilizando Markdown
-* As configurações são realizadas em um arquivo de formulário em PDF
-
-# Como iniciar utilização
-
-<!-- Show how to get started using or contribution to the project -->
-
-- [Veja se o visual do PDF gerado com a ferramenta lhe satisfaz](https://github.com/abntex/trabalho-academico-limarka/releases/download/v0.2.0/xxx-Monografia-projeto.pdf)
-- [Para começar a utilizar a ferramenta siga as instruções no wiki do projeto](https://github.com/abntex/limarka/wiki)
-
-# Contribuindo com o projeto
-
-<!-- State what resources and contributions you're looking for -->
-
-Veja como é fácil contribuir! 
-
-## Professor
-
-- Atualize sua página pessoal/institucional adicionando uma [matéria ou link sobre o limarka](https://github.com/abntex/limarka/wiki/Imprensa)
-- Solicite incluir o limarka como uma alternativa nos materiais de instruções de produção de trabalhos de conclusão, ou nas disciplinas de Metodologia.
-- Ao solicitar relatórios de atividades aos alunos, sugira a utilização do limarka.
-- Divulgue um notícia sobre o limarka nas listas de professores ou alunos.
-- Recomende a veiculação de uma matéria sobre o limarka na agência de noticias de sua instituição.
[…]
