### 2017-01-15 cf89020
Melhoria do texto #114
