### 2017-01-15 b5e9349
Atualização do readme  Writing a better CONTRIBUTING and README #99
