+++ README.md
@@ -10,11 +10,7 @@ Olá! Bem vindo ao limarka!
 
 O limarka é uma ferramenta que possibilita seus usuários escreverem relatórios ou trabalhos de conclusão de curso (Monografia, TCC, Dissertação ou Tese) em arquivos somente texto e **produzir PDFs em conformidades com as Normas da ABNT**. A produção desse tipo de trabalho é um momento sensível na vida dos estudantes, que pode contribuir para evasão no ensino superior. O objetivo do limarka é liberar dos estudantes os esforços de formatação e adequação às Normas da ABNT, deixando-os concentrar-se no planejamento, execução e escrita de seus trabalhos.
 
-# Utilizando o limarka
-
-![logo](https://raw.githubusercontent.com/wiki/abntex/limarka/imagens/escrevendo-objetivos-em-markdown.gif)
-
-# Quem utiliza o limarka diz que ele é fácil e prazeroso, descrubra o porquê!
+# Quem utiliza o limarka diz que ele é fácil e prazeroso, descubra o porquê!
 
 <!-- Explain what makes your project special, useful, exciting! -->
 
@@ -25,12 +21,16 @@ O limarka é uma ferramenta que possibilita seus usuários escreverem relatório
 * Permite a escrita do trabalho em arquivos somente texto, utilizando Markdown
 * As configurações são realizadas em um arquivo de formulário em PDF
 
+## Exemplo de utilização do limarka
+
+![logo](https://raw.githubusercontent.com/wiki/abntex/limarka/imagens/escrevendo-objetivos-em-markdown.gif)
+
 # Como iniciar utilização
 
 <!-- Show how to get started using or contribution to the project -->
 
-- [Veja se o visual do PDF gerado com a ferramenta lhe satisfaz](https://github.com/abntex/trabalho-academico-limarka/releases/download/v0.2.0/xxx-Monografia-projeto.pdf)
 - [Para começar a utilizar a ferramenta siga as instruções no wiki do projeto](https://github.com/abntex/limarka/wiki)
+- Lembre-se de acessar o [chat no gitter](http://gitter.im/abntex/limarka) para tirar suas dúvidas sobre a ferramenta.
 
 # Contribuindo com o projeto
 
@@ -60,12 +60,21 @@ Veja como é fácil contribuir!
 
 - Auxilie na criação de templates latex customizados para as instituições
 
-## Desenvolvedor Ruby 
+## Desenvolvedor Ruby
 
 [![GitHub issues](https://img.shields.io/github/issues/abntex/limarka.svg)](https://github.com/abntex/limarka/issues)
 
 - [Escolha uma das funcionalidades para implementar](https://github.com/abntex/limarka/issues)
 - Notifique ou corrija um Bug
+
+Acesse o [chat no gitter](http://gitter.im/abntex/limarka) e fale conosco!
+
+## Desenvolvedor Web
+
+Nós desejamos melhorar a experiência de utilização da ferramenta implementando 
+um site que gere conteúdos dinâmicos, gostaria de contribuir? 
+
+Acesse o [chat no gitter](http://gitter.im/abntex/limarka) e fale conosco!
 
 ## Todas as formas de contribuição
 
@@ -75,7 +84,7 @@ Consulte o arquivo [CONTRIBUTING.md](CONTRIBUTING.md) para verificar todas as fo
 
 ## Deseja entrar em contato?
 
-O meio de comunicação oficial do limarka é uma sala de bate papo no gitter: [![Join the chat at https://gitter.im/abntex/limarka](https://badges.gitter.im/abntex/limarka.svg)](https://gitter.im/abntex/limarka?utm_source=badge&utm_medium=badge&utm_campaign=pr-badge&utm_content=badge)
[…]
