### 2017-01-15 f266f13
Software paper writing closes #94
