### 2017-01-16 d378948
Updating version software of the paper #93 and #120
