+++ lib/limarka/version.rb
@@ -1,3 +1,3 @@
 module Limarka
-  VERSION = "0.6.3"
+  VERSION = "0.6.4"
 end
