### 2017-01-16 #136 @edusantana
@acabunoc I have translated the README.md and CONTRIBUTING.md files to english. You can see them at `v0.6.3-english` branch: https://github.com/abntex/limarka/tree/v0.6.3-english

The paper was also included in this version of code. It's DOI is https://doi.org/10.5281/zenodo.247247. The version of the software is now `v0.6.4` instead of `v0.5.2` and described on the issue description.

You can ask whatever you need for the review.
