### 2017-01-16 dcc4872
Atualizando documentação #121
