+++ lib/limarka/version.rb
@@ -1,3 +1,3 @@
 module Limarka
-  VERSION = "0.6.4"
+  VERSION = "0.6.5"
 end
