### 2017-01-16 #136 @edusantana
It's now `v0.6.5`, with API documentation.
DOI: https://doi.org/10.5281/zenodo.247387
