### 2017-01-22 27d0e54
Erro nas referências quando title está entre duplas chaves e contem subtítulo #122
Inclusão de Teste falhando #122
Fazendo teste passar #122

closes #122
