### 2017-01-22 5075819
Erro ao ler campos multi-linhas de configuração salva com o PDF-XChange Viewer
closes #124
