### 2017-01-22 dfb67d1
Atualizando versão para v0.6.6
