### 2017-01-23 #136 @edusantana
@acabunoc I'm using wiki as documentation, if you need that in English as well, I can make it happen, but takes time. Do you need any section translated?
