### 2017-01-24 #136 @acabunoc
Hi @edusantana! I think that will depend on the reviewer  -- I'll spend some time later to see if I find some other possible reviewers.  Thanks for all your work here! Let me know if you think of anyone else (doesn't have to be listed as a current reviewer for JOSS)
