### 2017-01-24 #136 @alanlivio
Hello @acabunoc and @edusantana.
I'm interested in review the Eduardo's paper.
I'm a PhD student at Puc-Rio Brazil and my research interests are in multimedia authoring and documents.
How I proceed to review?
