### 2017-01-24 #136 @acabunoc
Thanks @alanlivio! 🎉  I'll assign you as reviewer and get the review started. Review guidelines can be found here: http://joss.theoj.org/about#reviewer_guidelines

Do you read Portuguese? If so, @edusantana may not have to translate as much 😄
