### 2017-01-24 #169 @whedon
[REVIEW]: Limarka: making possible Brazilian students writing dissertation and thesis with Markdown
**Submitting author:** @edusantana (<a href="http://orcid.org/0000-0002-9165-9438">Eduardo Alexandre</a>)
**Repository:** <a href="https://github.com/abntex/limarka" target ="_blank">https://github.com/abntex/limarka</a>
**Version:** v0.6.6
