### 2017-01-24 #169 @acabunoc
version updated to v0.6.6
