### 2017-01-24 #169 @alanlivio
Hello @acabunoc.
I finish the review. But I found a problem in the form.
I believe that the item " Version: Does the release version given match the GitHub release (v0.5.2)?" should be changed to v0.6.6, right?
