### 2017-01-26 #169 @acabunoc
Also, @arfon, I manually updated the version number. Not sure if this breaks something behind the scenes.
