### 2017-01-26 #169 @edusantana
@acabunoc about the open canvas, I enjoy doing it. Thank you for that. I discovered it looking at your repositories. But [this tip](https://blog.cwrichardkim.com/how-to-get-hundreds-of-stars-on-your-github-project-345b065e20a2#.k7ckcoxgt) was also good, for creating an animated gif.
